refactor: move GlobalStyle to themes folder
