+++ src/themes/index.ts
@@ -1,2 +1,3 @@
 export * from './lightTheme';
 export * from './darkTheme';
+export * from './GlobalStyle';
